tests: Split DeleteMessageTest from test_message_edit.py.
